modbus-serial: introduce idleParser
Messages received from the serial line will be probably buffered
according to the FIFO size (8 bytes) which breaks the current
implementation, expecting the whole modbus RTU at once.
We introduce a silence-based parser for the serial module, which
allows to only emit the message when the line has been idle for a
given time (20ms -- found empirically).

A proper, more-reliable implementation would involve writing an
appropriate modbus protocol parser (including timeouts) but we'll
start with this proof-of-concept for now.

I tested this on a real system and it seems to work reliably for
reading holding registers sizes of up to 30 registers (resulting
in a well-formed PDU size of 65 bytes).
The system in use probably has some limited HW buffering because
everything starts to fail with bigger values, with 31 registers
resulting in a 65-byte buffer (missing the CRC).