Data reception with intermediate buffer
The use of an intermediate buffer is necessary in case onData functions fires multiple events during data reception instead of a single one as it was assumed before. To detect when a data reception is finished the onData function is fed by the onSend function with the number of bytes expected.