Fixed reqId calculation
Previously the reqId was incremented without modulo and could exceed the
value of 0xffff which would eventually throw an `out of range` error
when executing `Buffer.writeUInt16BE`.

Now, the reqId is incremented with a modulus on 0xffff and thus should
always fit into the word without and error. A test was added to verify
(and fix) the behavior.