Added sanity checks on requestId and protocolVersion in TCP Client
Until now, the only packet validation by the TCP Client was checking the
return function code. However, if an invalid function code is returned,
only the command promise is rejected but no consequences for the socket
are being made.
In this commit, I added sanity checks on the returned modbus header,
meaning that the expected requestId as well as the protocolVersion
(0x0000) are validated on every packet. If either of these is invalid,
an error is thrown and the socket is destroyed immediately.