init commit for UserRequestError
this creates a common error response class,
so that if an error is thrown or an exception
is occurs from a modbus request, the
expected class can be type checked and
dealt with when it arrises. For example,
in a try catch block, the function
`isUserRequestError` can be called to check
if the error occurred from the UserRequest.
e.g. `try{ }catch(err){isUserRequestError(err)}`