added ability to manually clear client requests
this is useful in the case where there is a modbus gateway
that is hosting multiple slave rtu devices and the request
is not technically timing out with no response, since the
TCP connection is still alive. This allows the client to flush
a number of requests if the cue is getting too large.